Separated logger and recipe base classes
